refactor: trim xi down to basics
