+++ packages/xi/constants.ts
@@ -1,11 +1,3 @@
-export const HANDLER_METHODS = new Set([
-  'GET', 'POST', 'PUT', 'PATCH', 'DELETE', 'HEAD', 'OPTIONS', 'FALLBACK'
-])
-
-export const HOOK_METHODS = new Set([
-  ...HANDLER_METHODS, 'ALL'
-])
-
 export const CONFIG = {
   base_path: ''
 }
