+++ packages/xin/lib/constants.js
@@ -5,3 +5,11 @@ export const DISALLOWED_METHODS = new Set([
 export const CONFIG = {
   check_origin: true 
 }
+
+export const HANDLER_METHODS = new Set([
+  'GET', 'POST', 'PUT', 'PATCH', 'DELETE', 'HEAD', 'OPTIONS', 'FALLBACK'
+])
+
+export const HOOK_METHODS = new Set([
+  ...HANDLER_METHODS, 'ALL'
+])
